Conflicts:
	data/shaders/opengl/geosphere_terrain.vert
	src/GeoPatch.cpp
	src/graphics/opengl/GeoSphereMaterial.cpp

+++ data/libs/SpaceStation.lua
@@ -1,4 +1,4 @@
--- Copyright © 2008-2014 Pioneer Developers. See AUTHORS.txt for details
+-- Copyright © 2008-2015 Pioneer Developers. See AUTHORS.txt for details
 -- Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
 local SpaceStation = import_core("SpaceStation")
@@ -204,39 +204,76 @@ local isPlayerShip = function (def) return def.tag == "SHIP" and def.basePrice >
 local groundShips = utils.build_array(utils.filter(function (k,def) return isPlayerShip(def) and def.equipSlotCapacity.atmo_shield > 0 end, pairs(ShipDef)))
 local spaceShips  = utils.build_array(utils.filter(function (k,def) return isPlayerShip(def) end, pairs(ShipDef)))
 
+
+-- Dynamics of ship adverts in ShipMarket --
+--------------------------------------------
+-- N(t) = Number of ads, lambda = decay constant:
+--    d/dt N(t) = prod - lambda * N
+-- and equilibrium:
+--    dN/dt = 0 = prod - lambda * N_equil
+-- and solution (if prod=0), with N_0 initial number:
+--    N(t) = N_0 * exp(-lambda * t)
+-- with tau = half life, i.e. N(tau) = 0.5*N_0 we get:
+--    0.5*N_0 = N_0*exp(-lambda * tau)
+-- else, with production:
+--   N(t) = prod/lambda - N_0*exp(-lambda * t)
+-- We want to have N_0 = N_equil, since ShipMarket should spawn in equilibrium
+
+-- Average number of ship for sale for station
+local function N_equilibrium(station)
+	local pop = station.path:GetSystemBody().parent.population -- E.g. Earth=7, Mars=0.3
+	local pop_bonus = 9 * math.log(pop*0.45 + 1)       -- something that gives resonable result
+	if station.type == "STARPORT_SURFACE" then
+		pop_bonus = pop_bonus * 1.5
+	end
+
+	return 2 + pop_bonus
+end
+
+-- add num random ships for sale to station ShipMarket
+local function addRandomShipAdvert(station, num)
+	for i=1,num do
+		local avail = station.type == "STARPORT_SURFACE" and groundShips or spaceShips
+		local def = avail[Engine.rand:Integer(1,#avail)]
+		local model = Engine.GetModel(def.modelName)
+		local pattern = model.numPatterns ~= 0 and Engine.rand:Integer(1,model.numPatterns) or nil
+		addShipOnSale(station, {
+			def     = def,
+			skin    = ModelSkin.New():SetRandomColors(Engine.rand):SetDecal(def.manufacturer),
+			pattern = pattern,
+			label   = Ship.MakeRandomLabel(),
+		})
+	end
+end
+
+local function createShipMarket (station)
+	shipsOnSale[station] = {}
+
+	local shipAdsToSpawn = Engine.rand:Poisson(N_equilibrium(station))
+	addRandomShipAdvert(station, shipAdsToSpawn)
+end
+
 local function updateShipsOnSale (station)
 	if not shipsOnSale[station] then shipsOnSale[station] = {} end
 	local shipsOnSale = shipsOnSale[station]
 
-	local toAdd, toRemove = 0, 0
-	if #shipsOnSale == 0 then
-		toAdd = Engine.rand:Integer(20)
-	elseif Engine.rand:Integer(1) > 0 then
-		toAdd = 1
-	elseif #shipsOnSale > 0 then
-		toRemove = 1
-	else
-		return
-	end
+	local tau = 7*24                              -- half life of a ship advert in hours
+	local lambda = 0.693147 / tau                 -- removal probability= ln(2) / tau
+	local prod = N_equilibrium(station) * lambda  -- creation probability
 
-	if toAdd > 0 then
-		local avail = station.type == "STARPORT_SURFACE" and groundShips or spaceShips
-		for i=1,toAdd do
-			local def = avail[Engine.rand:Integer(1,#avail)]
-			local model = Engine.GetModel(def.modelName)
-			local pattern = model.numPatterns ~= 0 and Engine.rand:Integer(1,model.numPatterns) or nil
-			addShipOnSale(station, {
-				def     = def,
-				skin    = ModelSkin.New():SetRandomColors(Engine.rand):SetDecal(def.manufacturer),
-				pattern = pattern,
-				label   = Ship.MakeRandomLabel(),
-			})
+	-- remove with decay rate lambda. Call ONCE/hour for each ship advert in station
+	for ref,ad in pairs(shipsOnSale) do
+		if Engine.rand:Number(0,1) < lambda then  -- remove one random ship (sold)
+			removeShipOnSale(station, Engine.rand:Integer(1,#shipsOnSale))
 		end
 	end
 
-	if toRemove > 0 then
-		removeShipOnSale(station, Engine.rand:Integer(1,#shipsOnSale))
+	-- spawn a new ship adverts, call for each station
+	if Engine.rand:Number(0,1) <= prod then
+		addRandomShipAdvert(station, 1)
 	end
+
+	if prod > 1 then print("Warning: ShipMarket not in equilibrium") end
 
 	Event.Queue("onShipMarketUpdate", station, shipsOnSale)
 end
@@ -246,7 +283,9 @@ end
 -- Group: Methods
 --
 
-
+SpaceStation.lockedAdvert = nil
+SpaceStation.advertLockCount = 0
+SpaceStation.removeOnReleased = false
 SpaceStation.adverts = {}
 
 --
@@ -363,9 +402,12 @@ end
 --
 --  stable
 --
-
 function SpaceStation:RemoveAdvert (ref)
 	if not SpaceStation.adverts[self] then return end
+	if SpaceStation.lockedAdvert == ref then
+		SpaceStation.removeOnReleased = true
+		return
+	end
 	local onDelete = SpaceStation.adverts[self][ref].onDelete
 	if onDelete then
 		onDelete(ref)
@@ -374,6 +416,42 @@ function SpaceStation:RemoveAdvert (ref)
 	Event.Queue("onAdvertRemoved", self, ref)
 end
 
+--
+-- Method: LockAdvert
+--
+-- > station:LockAdvert(ref)
+--
+-- Only one advert may be locked at a time. Locked adverts are not removed by
+-- RemoveAdvert until they are unlocked.
+--
+-- Parameters:
+--
+--   ref - the advert reference number returned by <AddAdvert>
+--
+-- Availability:
+--
+--  September 2014
+--
+-- Status:
+--
+--  experimental
+--
+function SpaceStation:LockAdvert (ref)
+	if (SpaceStation.advertLockCount > 0) then
+		assert(SpaceStation.lockedAdvert == ref, "Attempt to lock ref "..ref
+		.."disallowed."
+		.." Ref "..(SpaceStation.lockedAdvert or "nil").." is already locked "
+		..SpaceStation.advertLockCount.." times.")
+		SpaceStation.advertLockCount = SpaceStation.advertLockCount + 1
+		return
+	end
+	assert(SpaceStation.lockedAdvert == nil, "Attempt to lock ref "..ref
+		.." disallowed."
+		.." Ref "..(SpaceStation.lockedAdvert or "nil").." is already locked.")
+	SpaceStation.lockedAdvert = ref
+	SpaceStation.removeOnReleased = false
+	SpaceStation.advertLockCount = 1
+end
 
 local function updateAdverts (station)
 	-- XXX this should really just be a single event
@@ -386,6 +464,43 @@ local function updateAdverts (station)
 	end
 end
 
+--
+-- Method: UnlockAdvert
+--
+-- > station:UnlockAdvert(ref)
+--
+-- Releases the preserved advert. There must be an advert preserved at the
+-- time. If RemoveAdvert(ref) was called with the preserved advert's ref while
+-- the advert was preserved, it is now removed.
+--
+-- Parameters:
+--
+--   ref - the advert reference number returned by <AddAdvert>
+--
+-- Availability:
+--
+--  September 2014
+--
+-- Status:
+--
+--  experimental
+--
+function SpaceStation:UnlockAdvert (ref)
+	assert(SpaceStation.lockedAdvert == ref, "Attempt to unlock ref "..ref
+		.." disallowed."
+		.." Ref "..(SpaceStation.lockedAdvert or "nil").." is locked"
+		..SpaceStation.advertLockCount.." times. Unlock this"
+		.." first.")
+	if (SpaceStation.advertLockCount > 1) then
+		SpaceStation.advertLockCount = SpaceStation.advertLockCount - 1
+		return
+	end
+	SpaceStation.lockedAdvert = nil
+	SpaceStation.advertLockCount = 0
+	if SpaceStation.removeOnReleased then
+		self:RemoveAdvert(ref)
+	end
+end
 
 local function updateSystem ()
 	local stations = Space.GetBodies(function (b) return b.superType == "STARPORT" end)
@@ -395,6 +510,14 @@ local function updateSystem ()
 		updateAdverts(stations[i])
 	end
 end
+
+local function createSystem()
+	local stations = Space.GetBodies(function (b) return b.superType == "STARPORT" end)
+	for i=1,#stations do
+		createShipMarket(stations[i])
+	end
+end
+
 local function destroySystem ()
 	equipmentStock = {}
 	equipmentPrice = {}
@@ -433,11 +556,13 @@ Event.Register("onGameStart", function ()
 		loaded_data = nil
 	end
 
+	createSystem()
 	updateSystem()
 	Timer:CallEvery(3600, updateSystem)
 end)
 Event.Register("onEnterSystem", function (ship)
 	if ship ~= Game.player then return end
+	createSystem()
 	updateSystem()
 end)
